Auth refactor
Removed utils to composables
Move login part under layout
