--- composables/authState.js
@@ -1,9 +1,0 @@
-// It will be available as authState() (camelCase of file name without extension)
-export default function () {
-  return useState("authState", () => ({
-    user: null,
-    loading: false,
-    auth0: null,
-    isAuthenticated: false,
-  }));
-}
